Completed all my tasks and some other changes
*NonPlayer cannot take advantage of Give class(could not fix the issue
regarding albus targeting albus, see google Hangouts), so to enable give
for nonPlayer, a give method for nonplayer needs to be implemented. Code
is commented for the Give class implementation under albus for those who
want to have a go with it.

*Added a method that returns a boolean representing accept and Decline
in HPGridTextInterface and one other method to access this in
HPGridController

*Updated design rationale

*Other changes for Javadocs